backend/requirements.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
django-cors-headers==3.6.0